srvtopo: abstract more
