+++ go/vt/srvtopo/resilient_server.go
@@ -19,9 +19,6 @@ package srvtopo
 import (
 	"context"
 	"flag"
-	"fmt"
-	"html/template"
-	"sort"
 	"sync"
 	"time"
 
@@ -55,61 +52,6 @@ const (
 	queryCategory  = "query"
 	cachedCategory = "cached"
 	errorCategory  = "error"
-
-	// TopoTemplate is the HTML to use to display the
-	// ResilientServerCacheStatus object
-	TopoTemplate = `
-<style>
-  table {
-    border-collapse: collapse;
-  }
-  td, th {
-    border: 1px solid #999;
-    padding: 0.2rem;
-  }
-</style>
-<table>
-  <tr>
-    <th colspan="4">SrvKeyspace Names Cache</th>
-  </tr>
-  <tr>
-    <th>Cell</th>
-    <th>SrvKeyspace Names</th>
-    <th>TTL</th>
-    <th>Error</th>
-  </tr>
-  {{range $i, $skn := .SrvKeyspaceNames}}
-  <tr>
-    <td>{{github_com_vitessio_vitess_vtctld_srv_cell $skn.Cell}}</td>
-    <td>{{range $j, $value := $skn.Value}}{{github_com_vitessio_vitess_vtctld_srv_keyspace $skn.Cell $value}}&nbsp;{{end}}</td>
-    <td>{{github_com_vitessio_vitess_srvtopo_ttl_time $skn.ExpirationTime}}</td>
-    <td>{{if $skn.LastError}}({{github_com_vitessio_vitess_srvtopo_time_since $skn.LastQueryTime}}Ago) {{$skn.LastError}}{{end}}</td>
-  </tr>
-  {{end}}
-</table>
-<br>
-<table>
-  <tr>
-    <th colspan="5">SrvKeyspace Cache</th>
-  </tr>
-  <tr>
-    <th>Cell</th>
-    <th>Keyspace</th>
-    <th>SrvKeyspace</th>
-    <th>TTL</th>
-    <th>Error</th>
-  </tr>
-  {{range $i, $sk := .SrvKeyspaces}}
-  <tr>
-    <td>{{github_com_vitessio_vitess_vtctld_srv_cell $sk.Cell}}</td>
-    <td>{{github_com_vitessio_vitess_vtctld_srv_keyspace $sk.Cell $sk.Keyspace}}</td>
-    <td>{{$sk.StatusAsHTML}}</td>
-    <td>{{github_com_vitessio_vitess_srvtopo_ttl_time $sk.ExpirationTime}}</td>
-    <td>{{if $sk.LastError}}({{github_com_vitessio_vitess_srvtopo_time_since $sk.LastErrorTime}} Ago) {{$sk.LastError}}{{end}}</td>
-  </tr>
-  {{end}}
-</table>
-`
 )
 
 // ResilientServer is an implementation of srvtopo.Server based
@@ -117,18 +59,12 @@ const (
 // - limit the QPS to the underlying topo.Server
 // - return the last known value of the data if there is an error
 type ResilientServer struct {
-	topoServer   *topo.Server
-	cacheTTL     time.Duration
-	cacheRefresh time.Duration
-	counts       *stats.CountersWithSingleLabel
+	topoServer *topo.Server
+	counts     *stats.CountersWithSingleLabel
 
-	// mutex protects the cache map itself, not the individual
-	// values in the cache.
-	mutex                 sync.RWMutex
-	srvKeyspaceNamesCache map[string]*srvKeyspaceNamesEntry
-
-	srvKeyspaceWatcher *SrvKeyspaceWatcher
-	srvVSchemaWatcher  *SrvVSchemaWatcher
+	srvKeyspaceWatcher    *SrvKeyspaceWatcher
+	srvVSchemaWatcher     *SrvVSchemaWatcher
+	srvKeyspaceNamesQuery *SrvKeyspaceNamesQuery
 }
 
 type srvKeyspaceNamesEntry struct {
@@ -149,14 +85,6 @@ type srvKeyspaceNamesEntry struct {
 	lastErrorCtx  context.Context
 }
 
-type watchState int
-
-const (
-	watchStateIdle watchState = iota
-	watchStateStarting
-	watchStateRunning
-)
-
 // NewResilientServer creates a new ResilientServer
 // based on the provided topo.Server.
 func NewResilientServer(base *topo.Server, counterPrefix string) *ResilientServer {
@@ -165,7 +93,6 @@ func NewResilientServer(base *topo.Server, counterPrefix string) *ResilientServe
 	}
 
 	var metric string
-
 	if counterPrefix == "" {
 		metric = counterPrefix + "Counts"
 	} else {
@@ -174,14 +101,11 @@ func NewResilientServer(base *topo.Server, counterPrefix string) *ResilientServe
 	counts := stats.NewCountersWithSingleLabel(metric, "Resilient srvtopo server operations", "type")
 
 	return &ResilientServer{
-		topoServer:   base,
-		cacheTTL:     *srvTopoCacheTTL,
-		cacheRefresh: *srvTopoCacheRefresh,
-		counts:       counts,
-
-		srvKeyspaceNamesCache: make(map[string]*srvKeyspaceNamesEntry),
+		topoServer:            base,
+		counts:                counts,
 		srvKeyspaceWatcher:    NewSrvKeyspaceWatcher(base, counts, *srvTopoCacheRefresh, *srvTopoCacheTTL),
 		srvVSchemaWatcher:     NewSrvVSchemaWatcher(base, counts, *srvTopoCacheRefresh, *srvTopoCacheTTL),
+		srvKeyspaceNamesQuery: NewSrvKeyspaceNamesQuery(base, counts, *srvTopoCacheRefresh, *srvTopoCacheTTL),
 	}
 }
 
@@ -192,121 +116,7 @@ func (server *ResilientServer) GetTopoServer() (*topo.Server, error) {
 
 // GetSrvKeyspaceNames returns all keyspace names for the given cell.
 func (server *ResilientServer) GetSrvKeyspaceNames(ctx context.Context, cell string, staleOK bool) ([]string, error) {
-	server.counts.Add(queryCategory, 1)
-
-	// find the entry in the cache, add it if not there
-	key := cell
-	server.mutex.Lock()
-	entry, ok := server.srvKeyspaceNamesCache[key]
-	if !ok {
-		entry = &srvKeyspaceNamesEntry{
-			cell: cell,
-		}
-		server.srvKeyspaceNamesCache[key] = entry
-	}
-	server.mutex.Unlock()
-
-	// Lock the entry, and do everything holding the lock except
-	// querying the underlying topo server.
-	//
-	// This means that even if the topo server is very slow, two concurrent
-	// requests will only issue one underlying query.
-	entry.mutex.Lock()
-	defer entry.mutex.Unlock()
-
-	cacheValid := entry.value != nil && (time.Since(entry.insertionTime) < server.cacheTTL)
-	if !cacheValid && staleOK {
-		// Only allow stale results for a bounded period
-		cacheValid = entry.value != nil && (time.Since(entry.insertionTime) < (server.cacheTTL + 2*server.cacheRefresh))
-	}
-	shouldRefresh := time.Since(entry.lastQueryTime) > server.cacheRefresh
-
-	// If it is not time to check again, then return either the cached
-	// value or the cached error but don't ask topo again.
-	if !shouldRefresh {
-		if cacheValid {
-			return entry.value, nil
-		}
-		return nil, entry.lastError
-	}
-
-	// Refresh the state in a background goroutine if no refresh is already
-	// in progress none is already running. This way queries are not blocked
-	// while the cache is still valid but past the refresh time, and avoids
-	// calling out to the topo service while the lock is held.
-	if entry.refreshingChan == nil {
-		entry.refreshingChan = make(chan struct{})
-		entry.lastQueryTime = time.Now()
-		go func() {
-			defer func() {
-				if err := recover(); err != nil {
-					log.Errorf("GetSrvKeyspaceNames uncaught panic, cell :%v, err :%v)", cell, err)
-				}
-			}()
-			newCtx, cancel := context.WithTimeout(ctx, *srvTopoTimeout)
-			defer cancel()
-			result, err := server.topoServer.GetSrvKeyspaceNames(newCtx, cell)
-			entry.mutex.Lock()
-			defer func() {
-				close(entry.refreshingChan)
-				entry.refreshingChan = nil
-				entry.mutex.Unlock()
-			}()
-
-			if err == nil {
-				// save the value we got and the current time in the cache
-				entry.insertionTime = time.Now()
-				// Avoid a tiny race if TTL == refresh time (the default)
-				entry.lastQueryTime = entry.insertionTime
-				entry.value = result
-			} else {
-				server.counts.Add(errorCategory, 1)
-				if entry.insertionTime.IsZero() {
-					log.Errorf("GetSrvKeyspaceNames(%v, %v) failed: %v (no cached value, caching and returning error)", ctx, cell, err)
-				} else if newCtx.Err() == context.DeadlineExceeded {
-					log.Errorf("GetSrvKeyspaceNames(%v, %v) failed: %v (request timeout), (keeping cached value: %v)", ctx, cell, err, entry.value)
-				} else if entry.value != nil && time.Since(entry.insertionTime) < server.cacheTTL {
-					server.counts.Add(cachedCategory, 1)
-					log.Warningf("GetSrvKeyspaceNames(%v, %v) failed: %v (keeping cached value: %v)", ctx, cell, err, entry.value)
-				} else {
-					log.Errorf("GetSrvKeyspaceNames(%v, %v) failed: %v (cached value expired)", ctx, cell, err)
-					entry.insertionTime = time.Time{}
-					entry.value = nil
-				}
-			}
-
-			entry.lastError = err
-			entry.lastErrorCtx = newCtx
-		}()
-	}
-
-	// If the cached entry is still valid then use it, otherwise wait
-	// for the refresh attempt to complete to get a more up to date
-	// response.
-	//
-	// In the onEventLocked that the topo service is slow or unresponsive either
-	// on the initial fetch or if the cache TTL expires, then several
-	// requests could be blocked on refreshingCond waiting for the response
-	// to come back.
-	if cacheValid {
-		return entry.value, nil
-	}
-
-	refreshingChan := entry.refreshingChan
-	entry.mutex.Unlock()
-	select {
-	case <-refreshingChan:
-	case <-ctx.Done():
-		entry.mutex.Lock()
-		return nil, fmt.Errorf("timed out waiting for keyspace names")
-	}
-	entry.mutex.Lock()
-
-	if entry.value != nil {
-		return entry.value, nil
-	}
-
-	return nil, entry.lastError
+	return server.srvKeyspaceNamesQuery.Get(ctx, cell, staleOK)
 }
 
 // GetSrvKeyspace returns SrvKeyspace object for the given cell and keyspace.
@@ -317,152 +127,4 @@ func (server *ResilientServer) GetSrvKeyspace(ctx context.Context, cell, keyspac
 // WatchSrvVSchema is part of the srvtopo.Server interface.
 func (server *ResilientServer) WatchSrvVSchema(ctx context.Context, cell string, callback func(*vschemapb.SrvVSchema, error)) {
 	server.srvVSchemaWatcher.Watch(ctx, cell, callback)
-}
-
-// The next few structures and methods are used to get a displayable
-// version of the cache in a status page.
-
-// SrvKeyspaceNamesCacheStatus is the current value for SrvKeyspaceNames
-type SrvKeyspaceNamesCacheStatus struct {
-	Cell           string
-	Value          []string
-	ExpirationTime time.Time
-	LastQueryTime  time.Time
-	LastError      error
-	LastErrorCtx   context.Context
-}
-
-// SrvKeyspaceNamesCacheStatusList is used for sorting
-type SrvKeyspaceNamesCacheStatusList []*SrvKeyspaceNamesCacheStatus
-
-// Len is part of sort.Interface
-func (skncsl SrvKeyspaceNamesCacheStatusList) Len() int {
-	return len(skncsl)
-}
-
-// Less is part of sort.Interface
-func (skncsl SrvKeyspaceNamesCacheStatusList) Less(i, j int) bool {
-	return skncsl[i].Cell < skncsl[j].Cell
-}
-
-// Swap is part of sort.Interface
-func (skncsl SrvKeyspaceNamesCacheStatusList) Swap(i, j int) {
-	skncsl[i], skncsl[j] = skncsl[j], skncsl[i]
-}
-
-// SrvKeyspaceCacheStatus is the current value for a SrvKeyspace object
-type SrvKeyspaceCacheStatus struct {
-	Cell           string
-	Keyspace       string
-	Value          *topodatapb.SrvKeyspace
-	ExpirationTime time.Time
-	LastErrorTime  time.Time
-	LastError      error
-	LastErrorCtx   context.Context
-}
-
-// StatusAsHTML returns an HTML version of our status.
-// It works best if there is data in the cache.
-func (st *SrvKeyspaceCacheStatus) StatusAsHTML() template.HTML {
-	if st.Value == nil {
-		return template.HTML("No Data")
-	}
-
-	result := "<b>Partitions:</b><br>"
-	for _, keyspacePartition := range st.Value.Partitions {
-		result += "&nbsp;<b>" + keyspacePartition.ServedType.String() + ":</b>"
-		for _, shard := range keyspacePartition.ShardReferences {
-			result += "&nbsp;" + shard.Name
-		}
-		result += "<br>"
-	}
-
-	if st.Value.ShardingColumnName != "" {
-		result += "<b>ShardingColumnName:</b>&nbsp;" + st.Value.ShardingColumnName + "<br>"
-		result += "<b>ShardingColumnType:</b>&nbsp;" + st.Value.ShardingColumnType.String() + "<br>"
-	}
-
-	if len(st.Value.ServedFrom) > 0 {
-		result += "<b>ServedFrom:</b><br>"
-		for _, sf := range st.Value.ServedFrom {
-			result += "&nbsp;<b>" + sf.TabletType.String() + ":</b>&nbsp;" + sf.Keyspace + "<br>"
-		}
-	}
-
-	return template.HTML(result)
-}
-
-// SrvKeyspaceCacheStatusList is used for sorting
-type SrvKeyspaceCacheStatusList []*SrvKeyspaceCacheStatus
-
-// Len is part of sort.Interface
-func (skcsl SrvKeyspaceCacheStatusList) Len() int {
-	return len(skcsl)
-}
-
-// Less is part of sort.Interface
-func (skcsl SrvKeyspaceCacheStatusList) Less(i, j int) bool {
-	return skcsl[i].Cell+"."+skcsl[i].Keyspace <
-		skcsl[j].Cell+"."+skcsl[j].Keyspace
-}
-
-// Swap is part of sort.Interface
-func (skcsl SrvKeyspaceCacheStatusList) Swap(i, j int) {
-	skcsl[i], skcsl[j] = skcsl[j], skcsl[i]
-}
-
-// ResilientServerCacheStatus has the full status of the cache
-type ResilientServerCacheStatus struct {
-	SrvKeyspaceNames SrvKeyspaceNamesCacheStatusList
-	SrvKeyspaces     SrvKeyspaceCacheStatusList
-}
-
-// CacheStatus returns a displayable version of the cache
-func (server *ResilientServer) CacheStatus() *ResilientServerCacheStatus {
-	result := &ResilientServerCacheStatus{}
-
-	server.mutex.Lock()
-	for _, entry := range server.srvKeyspaceNamesCache {
-		entry.mutex.Lock()
-
-		result.SrvKeyspaceNames = append(result.SrvKeyspaceNames, &SrvKeyspaceNamesCacheStatus{
-			Cell:           entry.cell,
-			Value:          entry.value,
-			ExpirationTime: entry.insertionTime.Add(server.cacheTTL),
-			LastQueryTime:  entry.lastQueryTime,
-			LastError:      entry.lastError,
-			LastErrorCtx:   entry.lastErrorCtx,
-		})
-		entry.mutex.Unlock()
-	}
-	server.mutex.Unlock()
-
-	result.SrvKeyspaces = server.srvKeyspaceWatcher.CacheStatus()
-
-	// do the sorting without the mutex
-	sort.Sort(result.SrvKeyspaceNames)
-	sort.Sort(result.SrvKeyspaces)
-
-	return result
-}
-
-// Returns the ttl for the cached entry or "Expired" if it is in the past
-func ttlTime(expirationTime time.Time) template.HTML {
-	ttl := time.Until(expirationTime).Round(time.Second)
-	if ttl < 0 {
-		return template.HTML("<b>Expired</b>")
-	}
-	return template.HTML(ttl.String())
-}
-
-func timeSince(t time.Time) template.HTML {
-	return template.HTML(time.Since(t).Round(time.Second).String())
-}
-
-// StatusFuncs is required for CacheStatus) to work properly.
-// We don't register them inside servenv directly so we don't introduce
-// a dependency here.
-var StatusFuncs = template.FuncMap{
-	"github_com_vitessio_vitess_srvtopo_ttl_time":   ttlTime,
-	"github_com_vitessio_vitess_srvtopo_time_since": timeSince,
 }
